Add: script and support to identify who is using Plex.
Modifying the auth token script to have more correct and more specific
to Linux only (unfortantely). Adding the server users script so it is
linked with the auth token script so you can either specify a
username, password or a auth token.  Using the same auth token over
again is will obviously avoid Plex login messages on the server.