Fix: app_settings and env_info
app_settings was only partially implemented because it wasn't truely an
abstract settings store that could be used for any setting in any
script. That is now no longer the case. Other than 'verboose' which has
to be one of the settings its possible to set anything else as a setting
with anything as its value.

env_info got deleted one night as I was working while on sleeping pills.
Not a good idea! I believe now that its fully back to where it was
before I gutted the thing a week or two ago.

I'm also slowly going through each source file I touch and updating the
vim modeline so they are all identical.