Fixing and simplifying the read_physical_tty.py
The script now puts the burden of reading the pysical tty buffer on the
caller.  This is done because reading the physical tty often calls for
root access.  So instead of having to elevate the priveleges when
calling the script, we are now having the caller read the data prior
to executing.  Then that data can be piped into the script via `stdin`.

One valid example of calling `read_physical_tty.py`:

```bash
sudo cat /dev/vcsa | python3 read_physical_tty.py
```